requirements.txt
anthropic>=0.40.0
langgraph>=0.2.50
langchain-anthropic>=0.3.0
langchain-core>=0.3.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
sse-starlette>=2.1.0
pydantic>=2.10.0
